pyproject.toml
[tool.poetry]
name = "mastodon-covid19"
version = "0.1.0"
description = ""
authors = ["Mike Birdgeneau <mike.birdgeneau@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
"Mastodon.py" = "^1.5.2"
pandas = "^1.5.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
